refactor: appPrevoider, tag modules
