+++ libs/frontend/modules/app/src/provider/index.ts
@@ -1,4 +1,2 @@
 export * from './AppProvider'
-export * from './AppQueryProvider'
 export * from './types'
-export * from './useProvideCurrentApp'
