+++ libs/frontend/modules/page/src/providers/PageProvider.tsx
@@ -1,12 +1,14 @@
 import { ElementGraphProvider } from '@codelab/frontend/modules/element'
+import { withProvider } from '@codelab/frontend/presenter/container'
+import { useRouter } from 'next/router'
 import * as React from 'react'
-import { PageProviderProps } from './types'
 import { useProvideCurrentPage } from './useProvideCurrentPage'
 
-export const _PageProvider = ({
-  pageId,
+export const PageProvider = ({
   children,
-}: React.PropsWithChildren<PageProviderProps>) => {
+}: React.PropsWithChildren<unknown>) => {
+  const { query } = useRouter()
+  const pageId = query.pageId as string
   const { currentPage } = useProvideCurrentPage(pageId)
 
   return currentPage ? (
@@ -16,6 +18,4 @@ export const _PageProvider = ({
   ) : null
 }
 
-export const PageProvider = React.memo(_PageProvider, (prev, next) => {
-  return prev.pageId !== next.pageId
-})
+export const withPageProvider = withProvider(PageProvider)
