+++ libs/frontend/modules/page/src/providers/index.tsx
@@ -1,2 +1,1 @@
 export * from './PageProvider'
-export * from './PageQueryProvider'
