+++ libs/frontend/modules/page/src/use-cases/update-page/types.ts
@@ -1,8 +1,10 @@
 import { FormUniformsProps } from '@codelab/frontend/view/components'
 import { UpdatePageMutationVariables } from '../../graphql/Page.endpoints.graphql.gen'
 
-export type UpdatePageMutationInput =
-  UpdatePageMutationVariables['input']['updateData']
+export type UpdatePageMutationInput = Omit<
+  UpdatePageMutationVariables['input']['updateData'],
+  'rootElement'
+>
 
 export type UpdatePageFormProps = Omit<
   FormUniformsProps<UpdatePageMutationInput>,
