+++ libs/frontend/modules/tag/src/graphql/tag.endpoints.graphql
