+++ libs/frontend/modules/tag/src/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useTagState'
+export * from './useTagTree'
