+++ libs/frontend/modules/tag/src/hooks/useTagTree.tsx
@@ -1,6 +1,7 @@
 import { IGraph } from '@codelab/shared/abstract/core'
 import { TreeService } from '@codelab/shared/core'
-import { TagEdgeFragment, TagFragment } from '../../Tag.fragment.graphql.gen'
+import { TagFragment } from '../graphql/Tag.fragment.graphql.gen'
+import { TagEdgeFragment } from '../graphql/TagEdge.fragment.graphql.gen'
 
 export type GraphqlTagGraph = IGraph<TagFragment, TagEdgeFragment>
