+++ libs/frontend/modules/tag/src/index.ts
@@ -1,6 +1,4 @@
-export * from './hooks/use-tag/useTagState'
-export * from './store/tagEndpoints'
-export * from './store/tagState'
-export * from './Tag.fragment.graphql.gen'
-export * from './types/react-redux'
+export * from './hooks'
+export * from './store'
+export * from './types'
 export * from './use-cases'
