+++ libs/frontend/modules/tag/src/store/index.ts
@@ -1,0 +1,2 @@
+export * from './tagEndpoints'
+export * from './tagState'
